Refactor Course Catalog API to allow easily adding new endpoints.
With the requirement to add a Programs wrapper, we refactor
the Course Catalog API from Course Discovery to allow for more
easily dealing with fine-grained endpoints like Programs, Course Runs,
etc..
