+++ enterprise/admin/views.py
@@ -32,7 +32,7 @@ from enterprise.admin.utils import (
     split_usernames_and_emails,
     validate_email_to_link,
 )
-from enterprise.course_catalog_api import CourseCatalogApiClient
+from enterprise.course_discovery_api import CourseRunApiClient
 from enterprise.lms_api import EnrollmentApiClient, parse_lms_api_datetime
 from enterprise.models import (
     EnrollmentNotificationEmailTemplate,
@@ -498,7 +498,8 @@ class EnterpriseCustomerManageLearnersView(View):
             course_id: The specific course the learners were enrolled in
             users: An iterable of the users or pending users who were enrolled
         """
-        course_details = CourseCatalogApiClient(request.user).get_course_run(course_id)
+        course_run_api = CourseRunApiClient(request.user)
+        course_details = course_run_api.get_course_run(course_id)
         if not course_details:
             logging.warning(
                 _(
